fix new test
